+++ lib/pages/service_list_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class ServiceListScreen extends StatelessWidget {
+  const ServiceListScreen({Key? key, required this.label}) : super(key: key);
+  final String label;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        label,
+      ),
+    );
+  }
+}
